refactor(front): move the background boot pump out of App.svelte
The pump is its own concern; App.svelte only needs to start it and stop it on
destroy. No behaviour change — same staleness trigger, same sleep/wake
handover, same 60s ceiling — with the constants named and the reason the
interval stays armed while frames flow written down: a frame only proves the
window is visible right now, and a world can still be sent to the background
while it loads.
